+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Crowdloan tab shell
 - Gilt page shell
 - Society page shell
+- Democracy page shell
 
 #### Changed
 - Component i18n to global